Update B2B Workbench app to support both headless and B2B UI testing
