+++ source/sdk/src/main/AndroidManifest.xml
@@ -40,6 +40,23 @@
                     android:host="deeplink" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".ui.b2b.B2BAuthenticationActivity"
+            android:exported="true"
+            android:launchMode="singleTop"
+            android:screenOrientation="portrait"
+            android:windowSoftInputMode="adjustResize"
+            android:theme="@style/Theme.AppCompat.DayNight.NoActionBar">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW"/>
+                <category android:name="android.intent.category.DEFAULT"/>
+                <category android:name="android.intent.category.BROWSABLE"/>
+                <!-- These manifestPlaceholders in are defined in the application build.gradle -->
+                <data
+                    android:scheme="stytchui-${STYTCH_B2B_PUBLIC_TOKEN}"
+                    android:host="deeplink" />
+            </intent-filter>
+        </activity>
     </application>
     <queries>
         <intent>
