+++ workbench-apps/b2b-workbench/build.gradle
@@ -18,7 +18,8 @@ android {
         manifestPlaceholders = [
             'stytchOAuthRedirectScheme': 'app',
             'stytchOAuthRedirectHost': 'b2bOAuth',
-            'STYTCH_PUBLIC_TOKEN': rootProject.ext["STYTCH_PUBLIC_TOKEN"]
+            'STYTCH_PUBLIC_TOKEN': rootProject.ext["STYTCH_PUBLIC_TOKEN"],
+            'STYTCH_B2B_PUBLIC_TOKEN': rootProject.ext["STYTCH_B2B_PUBLIC_TOKEN"],
         ]
     }
